Refactoring
To get ready for work on #23

Work on #22
